requirements.txt
streamlit
numpy
opencv-python-headless
joblib
Pillow
scikit-learn